Split ModelEditorTest into focused test suites (#1449)
Extract CLD and subscript test groups from ModelEditorTest (2,686 lines)
into dedicated test files:

- ModelEditorCldTest: CLD variables, causal links, classification
- ModelEditorSubscriptTest: subscript preservation, assignment,
  definition management, copy-paste, element queries

ModelEditorTest drops from 2,686 to 2,039 lines. Each new file has
its own setUp() and focused scope for easier navigation and maintenance.
