+++ courant-app/src/test/java/systems/courant/sd/app/canvas/ModelEditorTest.java
@@ -1,8 +1,6 @@
 package systems.courant.sd.app.canvas;
 
 import systems.courant.sd.model.def.VariableDef;
-import systems.courant.sd.model.def.CausalLinkDef;
-import systems.courant.sd.model.def.CldVariableDef;
 import systems.courant.sd.model.def.ElementPlacement;
 import systems.courant.sd.model.def.ElementType;
 import systems.courant.sd.model.def.FlowDef;
@@ -12,7 +10,6 @@ import systems.courant.sd.model.def.ModelDefinitionBuilder;
 import systems.courant.sd.model.def.ModuleInstanceDef;
 import systems.courant.sd.model.def.SimulationSettings;
 import systems.courant.sd.model.def.StockDef;
-import systems.courant.sd.model.def.SubscriptDef;
 import systems.courant.sd.model.def.ViewDef;
 
 import java.util.ArrayList;
@@ -1871,374 +1868,6 @@ class ModelEditorTest {
         }
     }
 
-    @Nested
-    @DisplayName("CLD Variables")
-    class CldVariables {
-
-        @Test
-        void shouldAddCldVariable() {
-            String name = editor.addCldVariable();
-            assertThat(name).isEqualTo("Variable 1");
-            assertThat(editor.getCldVariables()).hasSize(1);
-            assertThat(editor.hasElement("Variable 1")).isTrue();
-        }
-
-        @Test
-        void shouldAutoIncrementCldVariableNames() {
-            editor.addCldVariable();
-            String second = editor.addCldVariable();
-            assertThat(second).isEqualTo("Variable 2");
-        }
-
-        @Test
-        void shouldRemoveCldVariable() {
-            editor.addCldVariable();
-            editor.removeElement("Variable 1");
-            assertThat(editor.getCldVariables()).isEmpty();
-            assertThat(editor.hasElement("Variable 1")).isFalse();
-        }
-
-        @Test
-        void shouldRenameCldVariable() {
-            editor.addCldVariable();
-            boolean result = editor.renameElement("Variable 1", "Workload");
-            assertThat(result).isTrue();
-            assertThat(editor.getCldVariables().get(0).name()).isEqualTo("Workload");
-            assertThat(editor.hasElement("Workload")).isTrue();
-            assertThat(editor.hasElement("Variable 1")).isFalse();
-        }
-
-        @Test
-        void shouldSetCldVariableComment() {
-            editor.addCldVariable();
-            boolean result = editor.setCldVariableComment("Variable 1", "A description");
-            assertThat(result).isTrue();
-            assertThat(editor.getCldVariables().get(0).comment()).isEqualTo("A description");
-        }
-
-        @Test
-        void shouldLoadCldVariablesFromDefinition() {
-            ModelDefinition def = new ModelDefinitionBuilder()
-                    .name("CLD")
-                    .cldVariable("Stress", "Work stress")
-                    .cldVariable("Performance")
-                    .causalLink("Stress", "Performance", CausalLinkDef.Polarity.NEGATIVE)
-                    .build();
-
-            editor.loadFrom(def);
-
-            assertThat(editor.getCldVariables()).hasSize(2);
-            assertThat(editor.getCausalLinks()).hasSize(1);
-            assertThat(editor.hasElement("Stress")).isTrue();
-            assertThat(editor.hasElement("Performance")).isTrue();
-        }
-
-        @Test
-        void shouldIncludeCldDataInSnapshot() {
-            editor.addCldVariable(); // Variable 1
-            editor.addCldVariable(); // Variable 2
-            editor.addCausalLink("Variable 1", "Variable 2", CausalLinkDef.Polarity.POSITIVE);
-
-            ModelDefinition snapshot = editor.toModelDefinition();
-            assertThat(snapshot.cldVariables()).hasSize(2);
-            assertThat(snapshot.causalLinks()).hasSize(1);
-            assertThat(snapshot.causalLinks().get(0).polarity())
-                    .isEqualTo(CausalLinkDef.Polarity.POSITIVE);
-        }
-
-        @Test
-        void shouldClearCldDataOnReload() {
-            editor.addCldVariable();
-            editor.addCausalLink("Variable 1", "Variable 1", CausalLinkDef.Polarity.POSITIVE);
-
-            editor.loadFrom(new ModelDefinitionBuilder().name("Empty").build());
-            assertThat(editor.getCldVariables()).isEmpty();
-            assertThat(editor.getCausalLinks()).isEmpty();
-        }
-
-        @Test
-        void shouldContinueCounterAfterLoad() {
-            ModelDefinition def = new ModelDefinitionBuilder()
-                    .name("CLD")
-                    .cldVariable("Variable 5")
-                    .build();
-            editor.loadFrom(def);
-
-            String next = editor.addCldVariable();
-            assertThat(next).isEqualTo("Variable 6");
-        }
-    }
-
-    @Nested
-    @DisplayName("Causal Links")
-    class CausalLinks {
-
-        @Test
-        void shouldAddCausalLink() {
-            editor.addCldVariable(); // Variable 1
-            editor.addCldVariable(); // Variable 2
-
-            boolean result = editor.addCausalLink("Variable 1", "Variable 2",
-                    CausalLinkDef.Polarity.POSITIVE);
-
-            assertThat(result).isTrue();
-            assertThat(editor.getCausalLinks()).hasSize(1);
-        }
-
-        @Test
-        void shouldRejectCausalLinkWithMissingEndpoint() {
-            editor.addCldVariable(); // Variable 1
-
-            boolean result = editor.addCausalLink("Variable 1", "Nonexistent",
-                    CausalLinkDef.Polarity.POSITIVE);
-
-            assertThat(result).isFalse();
-            assertThat(editor.getCausalLinks()).isEmpty();
-        }
-
-        @Test
-        void shouldRemoveCausalLink() {
-            editor.addCldVariable();
-            editor.addCldVariable();
-            editor.addCausalLink("Variable 1", "Variable 2", CausalLinkDef.Polarity.POSITIVE);
-
-            boolean result = editor.removeCausalLink("Variable 1", "Variable 2");
-
-            assertThat(result).isTrue();
-            assertThat(editor.getCausalLinks()).isEmpty();
-        }
-
-        @Test
-        void shouldSetCausalLinkPolarity() {
-            editor.addCldVariable();
-            editor.addCldVariable();
-            editor.addCausalLink("Variable 1", "Variable 2", CausalLinkDef.Polarity.UNKNOWN);
-
-            boolean result = editor.setCausalLinkPolarity("Variable 1", "Variable 2",
-                    CausalLinkDef.Polarity.NEGATIVE);
-
-            assertThat(result).isTrue();
-            assertThat(editor.getCausalLinks().get(0).polarity())
-                    .isEqualTo(CausalLinkDef.Polarity.NEGATIVE);
-        }
-
-        @Test
-        void shouldRemoveCausalLinksWhenVariableDeleted() {
-            editor.addCldVariable(); // Variable 1
-            editor.addCldVariable(); // Variable 2
-            editor.addCldVariable(); // Variable 3
-            editor.addCausalLink("Variable 1", "Variable 2", CausalLinkDef.Polarity.POSITIVE);
-            editor.addCausalLink("Variable 2", "Variable 3", CausalLinkDef.Polarity.NEGATIVE);
-            editor.addCausalLink("Variable 3", "Variable 1", CausalLinkDef.Polarity.POSITIVE);
-
-            editor.removeElement("Variable 2");
-
-            // Only the link from 3->1 should survive
-            assertThat(editor.getCausalLinks()).hasSize(1);
-            assertThat(editor.getCausalLinks().get(0).from()).isEqualTo("Variable 3");
-            assertThat(editor.getCausalLinks().get(0).to()).isEqualTo("Variable 1");
-        }
-
-        @Test
-        void shouldUpdateCausalLinksWhenVariableRenamed() {
-            editor.addCldVariable(); // Variable 1
-            editor.addCldVariable(); // Variable 2
-            editor.addCausalLink("Variable 1", "Variable 2", CausalLinkDef.Polarity.POSITIVE);
-
-            editor.renameElement("Variable 1", "Workload");
-
-            CausalLinkDef link = editor.getCausalLinks().get(0);
-            assertThat(link.from()).isEqualTo("Workload");
-            assertThat(link.to()).isEqualTo("Variable 2");
-        }
-
-        @Test
-        void shouldAllowCausalLinkBetweenCldAndSfElements() {
-            editor.addCldVariable(); // Variable 1
-            editor.addStock();       // Stock 1
-
-            boolean result = editor.addCausalLink("Variable 1", "Stock 1",
-                    CausalLinkDef.Polarity.POSITIVE);
-
-            assertThat(result).isTrue();
-            assertThat(editor.getCausalLinks()).hasSize(1);
-        }
-
-        @Test
-        void shouldRejectDuplicateCausalLink() {
-            editor.addCldVariable(); // Variable 1
-            editor.addCldVariable(); // Variable 2
-
-            boolean first = editor.addCausalLink("Variable 1", "Variable 2",
-                    CausalLinkDef.Polarity.POSITIVE);
-            boolean second = editor.addCausalLink("Variable 1", "Variable 2",
-                    CausalLinkDef.Polarity.NEGATIVE);
-
-            assertThat(first).isTrue();
-            assertThat(second).isFalse();
-            assertThat(editor.getCausalLinks()).hasSize(1);
-            assertThat(editor.getCausalLinks().getFirst().polarity())
-                    .isEqualTo(CausalLinkDef.Polarity.POSITIVE);
-        }
-
-        @Test
-        void shouldAllowReverseLinkBetweenSameVariables() {
-            editor.addCldVariable(); // Variable 1
-            editor.addCldVariable(); // Variable 2
-
-            boolean forward = editor.addCausalLink("Variable 1", "Variable 2",
-                    CausalLinkDef.Polarity.POSITIVE);
-            boolean reverse = editor.addCausalLink("Variable 2", "Variable 1",
-                    CausalLinkDef.Polarity.NEGATIVE);
-
-            assertThat(forward).isTrue();
-            assertThat(reverse).isTrue();
-            assertThat(editor.getCausalLinks()).hasSize(2);
-        }
-    }
-
-    @Nested
-    @DisplayName("classifyCldVariable")
-    class ClassifyCldVariable {
-
-        @Test
-        void shouldClassifyAsStock() {
-            editor.addCldVariable(); // Variable 1
-
-            boolean result = editor.classifyCldVariable("Variable 1", ElementType.STOCK);
-
-            assertThat(result).isTrue();
-            assertThat(editor.getCldVariables()).isEmpty();
-            assertThat(editor.getStocks()).hasSize(1);
-            StockDef stock = editor.getStocks().getFirst();
-            assertThat(stock.name()).isEqualTo("Variable 1");
-            assertThat(stock.initialValue()).isEqualTo(0);
-        }
-
-        @Test
-        void shouldClassifyAsFlow() {
-            editor.addCldVariable(); // Variable 1
-
-            boolean result = editor.classifyCldVariable("Variable 1", ElementType.FLOW);
-
-            assertThat(result).isTrue();
-            assertThat(editor.getCldVariables()).isEmpty();
-            assertThat(editor.getFlows()).hasSize(1);
-            FlowDef flow = editor.getFlows().getFirst();
-            assertThat(flow.name()).isEqualTo("Variable 1");
-        }
-
-        @Test
-        void shouldClassifyAsAux() {
-            editor.addCldVariable(); // Variable 1
-
-            boolean result = editor.classifyCldVariable("Variable 1", ElementType.AUX);
-
-            assertThat(result).isTrue();
-            assertThat(editor.getCldVariables()).isEmpty();
-            assertThat(editor.getVariables()).hasSize(1);
-            VariableDef v = editor.getVariables().getFirst();
-            assertThat(v.name()).isEqualTo("Variable 1");
-        }
-
-        @Test
-        void shouldClassifyAsLiteralAux() {
-            editor.addCldVariable(); // Variable 1
-
-            boolean result = editor.classifyCldVariable("Variable 1", ElementType.AUX);
-
-            assertThat(result).isTrue();
-            assertThat(editor.getCldVariables()).isEmpty();
-            assertThat(editor.getVariables()).hasSize(1);
-            VariableDef v = editor.getVariables().getFirst();
-            assertThat(v.name()).isEqualTo("Variable 1");
-        }
-
-        @Test
-        void shouldPreserveComment() {
-            editor.addCldVariable(); // Variable 1
-            editor.setCldVariableComment("Variable 1", "Important variable");
-
-            editor.classifyCldVariable("Variable 1", ElementType.STOCK);
-
-            StockDef stock = editor.getStocks().getFirst();
-            assertThat(stock.comment()).isEqualTo("Important variable");
-        }
-
-        @Test
-        void shouldRemoveCausalLinksInvolvingClassifiedVariable() {
-            editor.addCldVariable(); // Variable 1
-            editor.addCldVariable(); // Variable 2
-            editor.addCldVariable(); // Variable 3
-            editor.addCausalLink("Variable 1", "Variable 2",
-                    CausalLinkDef.Polarity.POSITIVE);
-            editor.addCausalLink("Variable 3", "Variable 1",
-                    CausalLinkDef.Polarity.NEGATIVE);
-            editor.addCausalLink("Variable 2", "Variable 3",
-                    CausalLinkDef.Polarity.POSITIVE);
-
-            editor.classifyCldVariable("Variable 1", ElementType.STOCK);
-
-            // Links involving Variable 1 should be removed
-            assertThat(editor.getCausalLinks()).hasSize(1);
-            assertThat(editor.getCausalLinks().getFirst().from()).isEqualTo("Variable 2");
-            assertThat(editor.getCausalLinks().getFirst().to()).isEqualTo("Variable 3");
-        }
-
-        @Test
-        void shouldReturnFalseForNonExistentVariable() {
-            boolean result = editor.classifyCldVariable("NoSuchVar", ElementType.STOCK);
-
-            assertThat(result).isFalse();
-        }
-
-        @Test
-        void shouldReturnFalseForUnsupportedTargetType() {
-            editor.addCldVariable(); // Variable 1
-
-            boolean result = editor.classifyCldVariable("Variable 1", ElementType.MODULE);
-
-            assertThat(result).isFalse();
-            // Variable should still be a CLD variable
-            assertThat(editor.getCldVariables()).hasSize(1);
-        }
-
-        @Test
-        void shouldPreserveNameInNameIndex() {
-            editor.addCldVariable(); // Variable 1
-
-            editor.classifyCldVariable("Variable 1", ElementType.STOCK);
-
-            // Name should still be registered (as a stock now)
-            assertThat(editor.hasElement("Variable 1")).isTrue();
-        }
-
-        @Test
-        void shouldPreserveOtherCldVariables() {
-            editor.addCldVariable(); // Variable 1
-            editor.addCldVariable(); // Variable 2
-
-            editor.classifyCldVariable("Variable 1", ElementType.AUX);
-
-            assertThat(editor.getCldVariables()).hasSize(1);
-            assertThat(editor.getCldVariables().getFirst().name()).isEqualTo("Variable 2");
-        }
-
-        @Test
-        void shouldAvoidNameCollisionWithExistingElements() {
-            // Manually create a variable named "Variable 1" to simulate collision
-            ModelDefinition def = new ModelDefinitionBuilder()
-                    .name("Test")
-                    .constant("Variable 1", 0, "u")
-                    .build();
-            editor.loadFrom(def);
-
-            String name = editor.addCldVariable();
-
-            assertThat(name).isNotEqualTo("Variable 1");
-            assertThat(editor.hasElement(name)).isTrue();
-        }
-    }
 
     @Nested
     @DisplayName("toModelDefinition thread safety")
@@ -2407,280 +2036,4 @@ class ModelEditorTest {
         }
     }
 
-    @Nested
-    @DisplayName("subscript preservation on stock setters")
-    class StockSubscriptPreservation {
-
-        @Test
-        void shouldPreserveSubscriptsOnSetInitialValue() {
-            ModelDefinition def = new ModelDefinitionBuilder()
-                    .name("Test")
-                    .subscript("Region", List.of("North", "South"))
-                    .stock("Pop", 100, "Person", List.of("Region"))
-                    .build();
-            editor.loadFrom(def);
-
-            editor.setStockInitialValue("Pop", 200);
-
-            StockDef updated = editor.getStockByName("Pop").orElseThrow();
-            assertThat(updated.initialValue()).isEqualTo(200);
-            assertThat(updated.subscripts()).containsExactly("Region");
-        }
-
-        @Test
-        void shouldPreserveSubscriptsOnSetUnit() {
-            ModelDefinition def = new ModelDefinitionBuilder()
-                    .name("Test")
-                    .subscript("Region", List.of("North", "South"))
-                    .stock("Pop", 100, "Person", List.of("Region"))
-                    .build();
-            editor.loadFrom(def);
-
-            editor.setStockUnit("Pop", "Animal");
-
-            StockDef updated = editor.getStockByName("Pop").orElseThrow();
-            assertThat(updated.unit()).isEqualTo("Animal");
-            assertThat(updated.subscripts()).containsExactly("Region");
-        }
-
-        @Test
-        void shouldPreserveSubscriptsOnSetNegativeValuePolicy() {
-            ModelDefinition def = new ModelDefinitionBuilder()
-                    .name("Test")
-                    .subscript("Region", List.of("North", "South"))
-                    .stock("Pop", 100, "Person", List.of("Region"))
-                    .build();
-            editor.loadFrom(def);
-
-            editor.setStockNegativeValuePolicy("Pop", "ALLOW");
-
-            StockDef updated = editor.getStockByName("Pop").orElseThrow();
-            assertThat(updated.negativeValuePolicy()).isEqualTo("ALLOW");
-            assertThat(updated.subscripts()).containsExactly("Region");
-        }
-
-        @Test
-        void shouldPreserveSubscriptsOnSetComment() {
-            ModelDefinition def = new ModelDefinitionBuilder()
-                    .name("Test")
-                    .subscript("Region", List.of("North", "South"))
-                    .stock("Pop", 100, "Person", List.of("Region"))
-                    .build();
-            editor.loadFrom(def);
-
-            editor.setStockComment("Pop", "A comment");
-
-            StockDef updated = editor.getStockByName("Pop").orElseThrow();
-            assertThat(updated.comment()).isEqualTo("A comment");
-            assertThat(updated.subscripts()).containsExactly("Region");
-        }
-    }
-
-    @Nested
-    @DisplayName("subscript assignment setters")
-    class SubscriptAssignment {
-
-        @Test
-        void shouldSetStockSubscripts() {
-            ModelDefinition def = new ModelDefinitionBuilder()
-                    .name("Test")
-                    .stock("S", 0, "units")
-                    .build();
-            editor.loadFrom(def);
-
-            boolean updated = editor.setStockSubscripts("S", List.of("Region"));
-
-            assertThat(updated).isTrue();
-            assertThat(editor.getStockByName("S").orElseThrow().subscripts())
-                    .containsExactly("Region");
-        }
-
-        @Test
-        void shouldSetFlowSubscripts() {
-            ModelDefinition def = new ModelDefinitionBuilder()
-                    .name("Test")
-                    .flow("F", "0", "Day", null, null)
-                    .build();
-            editor.loadFrom(def);
-
-            boolean updated = editor.setFlowSubscripts("F", List.of("Age"));
-
-            assertThat(updated).isTrue();
-            assertThat(editor.getFlowByName("F").orElseThrow().subscripts())
-                    .containsExactly("Age");
-        }
-
-        @Test
-        void shouldSetVariableSubscripts() {
-            ModelDefinition def = new ModelDefinitionBuilder()
-                    .name("Test")
-                    .variable("V", "0", "units")
-                    .build();
-            editor.loadFrom(def);
-
-            boolean updated = editor.setVariableSubscripts("V", List.of("Region", "Age"));
-
-            assertThat(updated).isTrue();
-            assertThat(editor.getVariableByName("V").orElseThrow().subscripts())
-                    .containsExactly("Region", "Age");
-        }
-
-        @Test
-        void shouldReturnFalseForNonexistentElement() {
-            editor.loadFrom(new ModelDefinitionBuilder().name("Test").build());
-
-            assertThat(editor.setStockSubscripts("Missing", List.of("X"))).isFalse();
-            assertThat(editor.setFlowSubscripts("Missing", List.of("X"))).isFalse();
-            assertThat(editor.setVariableSubscripts("Missing", List.of("X"))).isFalse();
-        }
-    }
-
-    @Nested
-    @DisplayName("subscript definition management")
-    class SubscriptDefinitionManagement {
-
-        @Test
-        void shouldAddSubscriptDef() {
-            editor.loadFrom(new ModelDefinitionBuilder().name("Test").build());
-
-            editor.addSubscript(new SubscriptDef("Region", List.of("N", "S")));
-
-            assertThat(editor.getSubscripts()).hasSize(1);
-            assertThat(editor.getSubscripts().get(0).name()).isEqualTo("Region");
-            assertThat(editor.getSubscripts().get(0).labels()).containsExactly("N", "S");
-        }
-
-        @Test
-        void shouldUpdateSubscriptDef() {
-            ModelDefinition def = new ModelDefinitionBuilder()
-                    .name("Test")
-                    .subscript("Region", List.of("N", "S"))
-                    .build();
-            editor.loadFrom(def);
-
-            boolean updated = editor.updateSubscript("Region",
-                    new SubscriptDef("Region", List.of("N", "S", "E", "W")));
-
-            assertThat(updated).isTrue();
-            assertThat(editor.getSubscripts().get(0).labels())
-                    .containsExactly("N", "S", "E", "W");
-        }
-
-        @Test
-        void shouldRemoveSubscriptDef() {
-            ModelDefinition def = new ModelDefinitionBuilder()
-                    .name("Test")
-                    .subscript("Region", List.of("N", "S"))
-                    .build();
-            editor.loadFrom(def);
-
-            boolean removed = editor.removeSubscript("Region");
-
-            assertThat(removed).isTrue();
-            assertThat(editor.getSubscripts()).isEmpty();
-        }
-
-        @Test
-        void shouldReturnFalseWhenRemovingNonexistentSubscript() {
-            editor.loadFrom(new ModelDefinitionBuilder().name("Test").build());
-
-            assertThat(editor.removeSubscript("Missing")).isFalse();
-        }
-
-        @Test
-        void shouldReturnFalseWhenUpdatingNonexistentSubscript() {
-            editor.loadFrom(new ModelDefinitionBuilder().name("Test").build());
-
-            assertThat(editor.updateSubscript("Missing",
-                    new SubscriptDef("Missing", List.of("A")))).isFalse();
-        }
-    }
-
-    @Nested
-    @DisplayName("copy-paste preserves subscripts")
-    class CopyPasteSubscripts {
-
-        @Test
-        void shouldPreserveSubscriptsOnAddStockFrom() {
-            editor.loadFrom(new ModelDefinitionBuilder().name("Test").build());
-            StockDef template = new StockDef("Pop", null, 100, "Person",
-                    null, List.of("Region"));
-
-            String name = editor.addStockFrom(template);
-
-            StockDef copy = editor.getStockByName(name).orElseThrow();
-            assertThat(copy.subscripts()).containsExactly("Region");
-        }
-
-        @Test
-        void shouldPreserveSubscriptsOnAddFlowFrom() {
-            editor.loadFrom(new ModelDefinitionBuilder().name("Test").build());
-            FlowDef template = new FlowDef("Migration", null, "0.1",
-                    "Year", "Person", null, null, List.of("Region"));
-
-            String name = editor.addFlowFrom(template, null, null);
-
-            FlowDef copy = editor.getFlowByName(name).orElseThrow();
-            assertThat(copy.subscripts()).containsExactly("Region");
-        }
-
-        @Test
-        void shouldPreserveSubscriptsOnAddVariableFrom() {
-            editor.loadFrom(new ModelDefinitionBuilder().name("Test").build());
-            VariableDef template = new VariableDef("Rate", null, "0.5",
-                    "1/Year", List.of("Region"));
-
-            String name = editor.addVariableFrom(template, template.equation());
-
-            VariableDef copy = editor.getVariableByName(name).orElseThrow();
-            assertThat(copy.subscripts()).containsExactly("Region");
-        }
-    }
-
-    @Nested
-    @DisplayName("getElementSubscripts")
-    class GetElementSubscripts {
-
-        @Test
-        void shouldReturnSubscriptsForSubscriptedStock() {
-            ModelDefinition def = new ModelDefinitionBuilder()
-                    .name("Test")
-                    .subscript("Region", List.of("N", "S"))
-                    .stock("Pop", 100, "Person", List.of("Region"))
-                    .build();
-            editor.loadFrom(def);
-
-            assertThat(editor.getElementSubscripts("Pop")).containsExactly("Region");
-        }
-
-        @Test
-        void shouldReturnSubscriptsForSubscriptedFlow() {
-            ModelDefinition def = new ModelDefinitionBuilder()
-                    .name("Test")
-                    .subscript("Region", List.of("N", "S"))
-                    .flow("Migration", "0", "Year", null, null, List.of("Region"))
-                    .build();
-            editor.loadFrom(def);
-
-            assertThat(editor.getElementSubscripts("Migration")).containsExactly("Region");
-        }
-
-        @Test
-        void shouldReturnEmptyListForScalarElement() {
-            ModelDefinition def = new ModelDefinitionBuilder()
-                    .name("Test")
-                    .stock("S", 0, "units")
-                    .build();
-            editor.loadFrom(def);
-
-            assertThat(editor.getElementSubscripts("S")).isEmpty();
-        }
-
-        @Test
-        void shouldReturnEmptyListForUnknownElement() {
-            editor.loadFrom(new ModelDefinitionBuilder().name("Test").build());
-
-            assertThat(editor.getElementSubscripts("Missing")).isEmpty();
-        }
-    }
 }
